requirement.txt
numpy==1.23.1
pandas==1.5.0
scikit-learn==1.1.2
pillow